LayoutTests/fast/dom/MicroData/propertynodelist-add-remove-itemref-tokens-expected.txt

Test to ensure that adding or removing itemref tokens must update its PropertyNodeList.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Created element of type: div
PASS PNL.length is 0
PASS PNL.length is 1
PASS PNL[0] is parent.firstChild
PASS PNL.length is 0
PASS !PNL[0] is true
PASS length0 is 0
PASS length1 is 1
PASS foo0 is parent.firstChild
PASS length2 is 0
PASS !foo1 is true
PASS successfullyParsed is true

TEST COMPLETE
